// packages/meta/src/types/metadata-interface.ts
interface Viewport extends ViewportLayout {
	/**
	 * The theme color for the document.
	 *
	 * @example
	 * ```tsx
	 * themeColor: "#000000"
	 * // Renders <meta name="theme-color" content="#000000" />
	 *
	 * themeColor: { media: "(prefers-color-scheme: dark)", color: "#000000" }
	 * // Renders <meta name="theme-color" media="(prefers-color-scheme: dark)" content="#000000" />
	 *
	 * themeColor: [
	 *  { media: "(prefers-color-scheme: dark)", color: "#000000" },
	 *  { media: "(prefers-color-scheme: light)", color: "#ffffff" }
	 * ]
	 * // Renders <meta name="theme-color" media="(prefers-color-scheme: dark)" content="#000000" />
	 * // Renders <meta name="theme-color" media="(prefers-color-scheme: light)" content="#ffffff" />
	 * ```
	 */
	themeColor?:
		| null
		| string
		| ThemeColorDescriptor
		| ThemeColorDescriptor[]
		| undefined;

	/**
	 * The color scheme for the document.
	 *
	 * @example
	 * ```tsx
	 * colorScheme: "dark"
	 * // Renders <meta name="color-scheme" content="dark" />
	 * ```
	 */
	colorScheme?: null | ColorSchemeEnum | undefined;
}

interface ResolvedViewport extends ViewportLayout {
	themeColor: null | ThemeColorDescriptor[];
	colorScheme: null | ColorSchemeEnum;
}